Standardize on Sodium-Universal
Problem: This module tries to straddle multiple cryptography libraries
with slightly different usage. This unpredictability is dangerous in a
cryptography library and hard to maintain.

Solution: Drop support for TweetNaCl and libsodium-wrappers to present a
single unified interface that we inherit from Sodium-Universal. This
should vastly reduce the complexity and incompatibility around the
Chloride module. Unfortunately, this has a *tiny* problem where the
Sodium-JavaScript implementation doesn't support the functions we need.
Oops.
