--- bindings.js
@@ -1,4 +1,0 @@
-var _require = require //fool browserify
-module.exports =
-  _require('sodium-chloride')(_require('sodium-native'))
-
