--- browser-small.js
@@ -1,2 +1,0 @@
-
-module.exports = require('sodium-browserify-tweetnacl')
