--- chloride-test/README.md
@@ -1,16 +1,0 @@
-# chloride-test
-
-test that a module provides the same functions
-as [chloride](https://github.com/dominictarr/chloride)
-via tests generated from [chloridedown](https://github.com/dominictarr/chloridedown)
-
-## example
-
-``` js
-require('chloride-test')(require('sodium-browserify'))
-```
-
-
-## License
-
-MIT
