--- chloride-test/test.js
@@ -1,6 +1,0 @@
-
-process.exit(
-require('./')(require('sodium-chloride')(require('sodium-native'))).fail
-)
-
-
