+++ package.json
@@ -8,12 +8,10 @@
   "browser": "browser.js",
   "dependencies": {
     "is-electron": "^2.2.0",
-    "sodium-browserify": "^1.2.7",
-    "sodium-browserify-tweetnacl": "^0.2.5",
-    "sodium-chloride": "^1.1.2"
+    "sodium-universal": "^3.0.0"
   },
-  "optionalDependencies": {
-    "sodium-native": "^3.0.0"
+  "devDependencies": {
+    "json-buffer": "^3.0.1"
   },
   "repository": {
     "type": "git",
@@ -35,10 +33,7 @@
       "url": "http://opensource.org/licenses/mit-license.php"
     }
   ],
-  "devDependencies": {
-    "chloride-test": "^1.2.4"
-  },
   "scripts": {
-    "test": "node test.js"
+    "test": "node test"
   }
 }
