--- small.js
@@ -1,11 +1,0 @@
-if (process.env.CHLORIDE_JS) {
-  module.exports = require('./browser-small')
-} else {
-  try {
-    module.exports = require('./bindings')
-  } catch (err) {
-    console.error('error loading sodium bindings:', err.message)
-    console.error('falling back to javascript version.')
-    module.exports = require('./browser-small')
-  }
-}
