--- sodium-chloride/README.md
@@ -1,12 +1,0 @@
-# sodium-chloride
-
-polyfil to adapt [sodium-friends](https://github.com/sodium-friends) api to [chloride](https://github.com/dominictarr/chloride)
-
-## TODO
-
-* `*box_open*` methods arn't passing yet.
-* sodium-friends needs a PR to add key conversions.
-
-## License
-
-MIT
