--- sodium-chloride/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "sodium-chloride",
-  "description": "polyfil from chloride to sodium-friends",
-  "version": "1.1.2",
-  "homepage": "https://github.com/dominictarr/sodium-chloride",
-  "repository": {
-    "type": "git",
-    "url": "git://github.com/dominictarr/sodium-chloride.git"
-  },
-  "devDependencies": {
-    "chloride-test": "^1.2.2",
-    "sodium-native": "^3.0.0"
-  },
-  "scripts": {
-    "test": "node test"
-  },
-  "author": "'Dominic Tarr' <dominic.tarr@gmail.com> (http://dominictarr.com)",
-  "license": "MIT"
-}
