--- sodium-chloride/test.js
@@ -1,2 +1,0 @@
-
-process.exit(require('chloride-test')(require('./')(require('sodium-native'))).fail)
